Do path better
I decided to stop using the previous value for $PATH when I do it. This
usually caused all kinds of problems with nested sessions. It appears to
work well enough with just defining the entire path in the zshenv and
results in much more predictable behavior.
I also am now doing it using the zsh array `path` instead of `PATH`.